+++ posting/editing/edit_image.py
@@ -9,7 +9,8 @@ def put_text_on_image(image: Image, text: str, font_type: str, font_size: int, f
     draw.text((10, 10), text, font=font, fill=font_color)
     return image
 
-new_image = put_text_on_image(Image.open("image/test_image.jpeg"), "안녕하세요", "MaruBuri-Bold.ttf", 150, "black")
-plt.imshow(new_image)
-plt.axis('off')  # Turn off axis
-plt.show()
+if __name__ == "__main__":
+    new_image = put_text_on_image(Image.open("../image/test_image.jpeg"), "안녕하세요", "../font/MaruBuri-Bold.ttf", 150, "black")
+    plt.imshow(new_image)
+    plt.axis('off')
+    plt.show()
